Add pingBackend to spin up backend early (#9)
Currently in the backend side, the host service will put the process to
sleep after a certain amount of time without traffic. So on the first
code run, if the backend is sleeping, it will take a long time since the
backend needs to be waken up.

To solve this, we are sending a request to the backend on the app
startup to wake it up beforehand.